created ActivityPlayerPanel extends PlayerPanel for AudioActivity
