+++ app/src/main/java/com/irateam/vkplayer/viewholders/PlayerPanel.java
@@ -12,7 +12,6 @@ import android.widget.TextView;
 
 import com.irateam.vkplayer.R;
 import com.irateam.vkplayer.activities.AudioActivity;
-import com.irateam.vkplayer.activities.ListActivity;
 import com.irateam.vkplayer.player.Player;
 import com.irateam.vkplayer.services.PlayerService;
 import com.vk.sdk.api.model.VKApiAudio;
@@ -21,7 +20,6 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLConnection;
-import java.text.SimpleDateFormat;
 import java.util.concurrent.ExecutionException;
 
 public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerProgressListener {
@@ -29,10 +27,6 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
     public View rootView;
     public TextView songName;
     public TextView author;
-    public TextView currentTime;
-    public TextView timeToFinish;
-    public TextView numberAudio;
-    public TextView sizeAudio;
 
     public ImageView repeat;
     public ImageView previous;
@@ -56,10 +50,6 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
         rootView = view;
         songName = (TextView) view.findViewById(R.id.player_panel_song_name);
         author = (TextView) view.findViewById(R.id.player_panel_author);
-        currentTime = (TextView) view.findViewById(R.id.player_panel_current_time);
-        timeToFinish = (TextView) view.findViewById(R.id.player_panel_time_remaining);
-        numberAudio = (TextView) view.findViewById(R.id.player_panel_count_audio);
-        sizeAudio = (TextView) view.findViewById(R.id.player_panel_audio_size);
 
         repeat = (ImageView) view.findViewById(R.id.player_panel_repeat);
         previous = (ImageView) view.findViewById(R.id.player_panel_previous);
@@ -82,16 +72,10 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
             public void onClick(View v) {
                 if (playerService.isPlaying()) {
                     playerService.pause();
-                    if (!audioActivity)
-                        playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_play_grey_18dp));
-                    else
-                        playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_play_grey_24dp));
+                    playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_play_grey_18dp));
                 } else {
                     playerService.resume();
-                    if (!audioActivity)
-                        playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_pause_grey_18dp));
-                    else
-                        playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_pause_grey_24dp));
+                    playPause.setImageDrawable(resources.getDrawable(R.drawable.ic_player_pause_grey_18dp));
                 }
             }
         });
@@ -145,18 +129,13 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
         headerLayout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (!audioActivity) {
-                    Intent intent = new Intent(context, AudioActivity.class);
-                    context.startActivity(intent);
-                } else {
-                    Intent intent = new Intent(context, ListActivity.class);
-                    context.startActivity(intent);
-                }
+                Intent intent = new Intent(context, AudioActivity.class);
+                context.startActivity(intent);
             }
         });
     }
 
-    private void configurePanel(PlayerService playerService) {
+    public void configurePanel(PlayerService playerService) {
         playerService.addPlayerProgressListener(this);
         VKApiAudio audio = playerService.getPlayingAudio();
         if (audio != null) {
@@ -186,45 +165,6 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
             progress.setMax(audio.duration * 1000);
             progress.setProgress(0);
             progress.setSecondaryProgress(0);
-            if (audioActivity) {
-                numberAudio.setText("#" + (position + 1) + "/");
-                try {
-                    SizeTask sizeTask = new SizeTask(new URL(audio.url));
-                    sizeTask.execute();
-                    String size = String.format("%.1f", Double.valueOf(sizeTask.get()));
-                    sizeAudio.setText(size + "Mb");
-                } catch (MalformedURLException e) {
-                    e.printStackTrace();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                } catch (ExecutionException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-    class SizeTask extends AsyncTask<Void, Void, Double> {
-
-        private URL url;
-
-        public SizeTask(URL url) {
-            this.url = url;
-        }
-
-        @Override
-        protected Double doInBackground(Void... params) {
-
-            URLConnection urlConnection = null;
-            double size = 0;
-            try {
-                urlConnection = url.openConnection();
-                urlConnection.connect();
-                size = urlConnection.getContentLength() / (double) 1024 / (double) 1024;
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            return size;
         }
     }
 
@@ -244,7 +184,7 @@ public class PlayerPanel implements Player.PlayerEventListener, Player.PlayerPro
     }
 
     @SuppressWarnings("deprecation")
-    private void setRandomState(boolean randomState) {
+    public void setRandomState(boolean randomState) {
         if (randomState) {
             random.setImageDrawable(resources.getDrawable(R.drawable.ic_player_random_on_light_grey_18dp));
         } else {
