+++ app/src/main/res/menu/menu_audio.xml
@@ -1,0 +1,4 @@
+<menu xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.irateam.vkplayer.activities.AudioActivity">
+
+</menu>
