+++ app/src/main/res/values-v21/styles.xml
@@ -1,0 +1,2 @@
+<resources>>
+</resources>
